+++ app/src/main/java/com/ikvych/cocktail/ui/activity/SearchActivity.java
@@ -5,8 +5,8 @@ import android.widget.SearchView;
 
 import com.ikvych.cocktail.R;
 import com.ikvych.cocktail.data.entity.Drink;
-import com.ikvych.cocktail.ect.adapter.DrinkAdapter;
-import com.ikvych.cocktail.ect.util.ActivityUtil;
+import com.ikvych.cocktail.adapter.list.DrinkAdapter;
+import com.ikvych.cocktail.util.ActivityUtil;
 import com.ikvych.cocktail.ui.base.BaseActivity;
 import com.ikvych.cocktail.viewmodel.SearchActivityViewModel;
